move PaymentMethodsList in a dedicated folder
